requirements.txt
python-telegram-bot
openai-whisper
torch
python-dotenv

langgraph
langchain
langchain-ollama

pandas
matplotlib
reportlab